requirements.txt
pandas~=2.0.2
numpy~=1.24.3
Pillow~=9.5.0
scikit-learn~=1.2.2
requests~=2.31.0
Flask~=2.3.2